requirements.txt
colorama==0.4.4
tqdm==4.62.3
cryptography==3.4.7
requests==2.26.0
networkx==2.6.3
matplotlib==3.4.3
textblob==0.15.3
opencv-python==4.5.3.56
Pillow==8.3.2
piexif==1.1.3
folium==0.12.1
geopy==2.2.0
instaloader==4.8.1